chore(project-structure): reorganizar proyecto a arquitectura en capas
- Mover AppController y AppService a modules/app

- Agregar PrismaService en carpeta prisma

- Crear carpetas comunes y config para futura expansión

- Preparar estructura para integración con Prisma y Docker
