+++ src/app.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { AppController } from './modules/app/controllers/app.controller';
+import { AppService } from './modules/app/services/app.service';
+import { PrismaService } from './prisma/prisma.service';
 
 @Module({
   imports: [],
   controllers: [AppController],
-  providers: [AppService],
+  providers: [AppService,PrismaService],
 })
 export class AppModule {}
